pyproject.toml
[project]
name = "agent-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docker>=7.1.0",
    "gradio>=4.19.2",
    "ollama>=0.2.1", # Added Ollama library
    "loguru>=0.7.3",
    "mcp[cli]>=1.3.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "pytz>=2025.2",
    "jsonschema>=4.23.0",
]
